src/backend/requirements.txt
fastapi
uvicorn
openai
PyPDF2
pillow
python-dotenv
python-multipart
camelot-py
opencv-python
pytesseract